+++ src/templates/seo-head.js
@@ -3,6 +3,7 @@ export const nuxtI18nSeo = function () {
   if (
     !this._hasMetaInfo ||
     !this.$i18n ||
+    !this.$i18n.locale ||
     !this.$i18n.locales ||
     this.$options[COMPONENT_OPTIONS_KEY] === false ||
     (this.$options[COMPONENT_OPTIONS_KEY] && this.$options[COMPONENT_OPTIONS_KEY].seo === false)